:recycle: Moving constants into fadeInText folder
