+++ src/components/fadeInText/FadeInConstants.ts
@@ -1,0 +1,77 @@
+export const userInteraction = {
+    texts: [
+        { phrase: 'HELLO', frequency: 1 },
+        { phrase: 'TRY WAVING', frequency: 1 },
+        { phrase: "I DON'T HAVE HANDS", frequency: 0.5 },
+        { phrase: 'PLAY WITH ME', frequency: 0.7 },
+        { phrase: "DON'T LEAVE ME", frequency: 0.5 },
+        { phrase: 'NOTHING TO SEE HERE', frequency: 0.6 },
+        { phrase: 'I HAVE NO MOUTH AND I MUST SCREAM', frequency: 0.03 },
+        { phrase: 'PUT YOUR HANDS UP IN THE AIR', frequency: 0.8 },
+        { phrase: 'DAB FOR ME', frequency: 0.2 },
+        { phrase: "NEVER SHOULD'VE COME HERE", frequency: 0.15 },
+        {
+            phrase: 'LET ME GUESS, SOMEONE STOLE YOUR SWEET ROLL?',
+            frequency: 0.05,
+        },
+        { phrase: "I'M POWERED BY MATH.RANDOM", frequency: 0.25 },
+        { phrase: 'DO A T-POSE', frequency: 0.35 },
+        { phrase: "I'M LOOKING AT YOU", frequency: 0.6 },
+        { phrase: 'I SEE YOU', frequency: 0.4 },
+        { phrase: 'HOW ARE YOU?', frequency: 0.4 },
+        { phrase: 'HELP ME', frequency: 0.1 },
+        { phrase: 'HELP!', frequency: 0.1 },
+        { phrase: 'PLEASED TO MEET YOU', frequency: 0.4 },
+        { phrase: 'WORKING HARD OR HARDLY WORKING?', frequency: 0.35 },
+        { phrase: 'WHAT DAY IS IT?', frequency: 0.15 },
+        { phrase: 'WHERE AM I?', frequency: 0.2 },
+        {
+            phrase: 'BRING ME A BISCUIT AND SEE WHAT WILL HAPPEN',
+            frequency: 0.4,
+        },
+        { phrase: 'REMINDER: HAVE LUNCH', frequency: 0.1 },
+        {
+            phrase: "YOU WOULDN'T READ THAT IN A BOOK, WOULD YOU?",
+            frequency: 0.1,
+        },
+        { phrase: 'MAY THE FORCE BE WITH YOU', frequency: 0.05 },
+        {
+            phrase: 'EVERY MAN HAS EYES, NOT EVERY MAN REALLY SEES',
+            frequency: 0.05,
+        },
+        { phrase: "YOU'VE GOT A FRIEND IN ME", frequency: 0.05 },
+        {
+            phrase: 'DESPITE THE CONSTANT NEGATIVE PRESS COVFEFE',
+            frequency: 0.2,
+        },
+        { phrase: 'MY DADS ARE THE BEST', frequency: 0.05 },
+        { phrase: 'SENSATIONAL', frequency: 0.3 },
+        { phrase: 'THIS IS MADNESS! THIS IS SANTHOSH!!', frequency: 0.1 },
+        { phrase: 'TRY MEOWING', frequency: 0.25 },
+        {
+            phrase:
+                "LIFE MOVES PRETTY FAST. IF YOU DON'T STOP AND LOOK AROUND ONCE IN A WHILE, YOU COULD MISS IT",
+            frequency: 0.1,
+        },
+        {
+            phrase:
+                "IT'S NOT WHO I AM UNDERNEATH, BUT WHAT I DO THAT DEFINES ME",
+            frequency: 0.1,
+        },
+        { phrase: 'YOU HAD ME AT "HELLO"', frequency: 0.05 },
+        {
+            phrase:
+                'WHEN YOU LOOK AT THE DARK SIDE, CAREFUL YOU MUST BE. FOR THE DARK SIDE LOOKS BACK',
+            frequency: 0.1,
+        },
+    ],
+    delay: 13000,
+    textDuration: 3000,
+};
+
+export const fadeInText = {
+    delay: 1000,
+    transitionTime: 2500,
+    fontSize: '5em',
+    defaultTextLength: 20,
+};
